AbstractHeatmapCanvas class
